[PC-14755] refacto api reco (#749)
* [PC-14755] WIP

* [PC-14755] WIP

* [PC-14755]  WIP clean scoring methods

* [PC-14755] WIP

* [PC-14755] WIP cold start status

* [PC-14755-refacto-api-reco] WIP fix _get_instances

* [PC-14755] WIP fix env_vars naming

* [PC-14755] WIP fix user init

* [PC-14755] add save recommendation

* [PC-14755] fix save recommendation empty reco case

* [PC-14755] update postman integration tests

* [PC-14755] fix integration tests format

* [PC-14755] WIP

* [PC-14755] WIP

* [PC-14755] WIP fix tests

* [PC-14755] rm recommendation_tests.py

* [PC-14755] rm recommendation.py

* [PC-14755] clean integration tests

* [PC-14755] clean integration tests bis

* [PC-14755] fix tests unitaires

* [PC-14755] fix tests unitaires bis

* [PC-14755] fix tests unitaires ter

* [PC-14755] update code structure

* [PC-14755] update pytest.ini testpath

* [PC-14755] clean naming

* [PC-14755] update integration tests

* [PC-14755] WIP fix tests unitaires

* [PC-14755] fix relative path

* [PC-14755] add Flags for debug

* [PC-14755] bypass unitary tests
