+++ recommendation/api/postman/dev.postman_environment.json
@@ -3,13 +3,8 @@
 	"name": "Dev",
 	"values": [
 		{
-			"key": "user_id_A",
+			"key": "user_id",
 			"value": "1",
-			"enabled": true
-		},
-		{
-			"key": "user_id_B",
-			"value": "2",
 			"enabled": true
 		},
 		{
