+++ recommendation/api/pytest.ini
@@ -1,6 +1,6 @@
 [pytest]
 addopts=-v --tb=short
-testpaths=tests
+testpaths=src/tests
 norecursedirs=.git venv/ .pytest_cache/
 python_files=*_test.py test_*.py
 python_classes=Post Patch Put Get Delete Returns* When* *Test
